+++ composables/useCanvasControls.ts
@@ -2,7 +2,7 @@
 
 import { FabricImage, Textbox } from 'fabric'
 import type { Canvas, FabricObject } from 'fabric'
-import { CircularTextbox } from '~/utils/CircularTextbox'
+import { CircularTextbox } from '~/utils/circularTextbox'
 import { setTextboxTextRadius } from '~/utils/customDesign'
 import { useCustomImage } from '~/composables/useCustomImage'
 import { useCustomText } from '~/composables/useCustomText'
